jeefs_endian: avoid signed-shift overflow in jeefs_get_le16
p[1] promotes to signed int; with a 16-bit int, values >= 0x80 shifted
left by 8 overflow it — undefined behaviour. Cast to uint16_t before
the shift, matching the 32-bit helper.

+++ include/jeefs_endian.h
@@ -18,7 +18,7 @@ extern "C" {
 #endif
 
 static inline uint16_t jeefs_get_le16(const uint8_t *p) {
-    return (uint16_t)(p[0] | (p[1] << 8));
+    return (uint16_t)(p[0] | ((uint16_t)p[1] << 8));
 }
 
 static inline uint32_t jeefs_get_le32(const uint8_t *p) {
